refactor(achievements): consolidate /achievements routes, drop legacy goals shape
The unified-table migration (unification B–D) left the single
/achievements endpoint split across two route files with tangled
ownership: the PUT/GET handlers lived in +Achievements.swift but called
conversions defined in +UnifiedAchievements.swift, and the handlers
still accepted the legacy {goals:[...]} body alongside the unified
{achievements:[...]} shape.

No client sends the legacy shape anymore — achievements-editor.js reads
and writes only `achievements`, and no MCP tool touches the endpoint —
so this folds everything into PublishedAssignmentRoutes+Achievements.swift
and removes the legacy shape:

* AchievementsBody/AchievementsResponse now carry only `achievements`.
* ClassGoalInput, the goals-only merge branch, and the goals response
  projection are gone; every PUT replaces the whole list and marks
  builtInAchievementsSeeded.
* PublishedAssignmentRoutes+UnifiedAchievements.swift is deleted; the
  row type + conversions move into the one file that owns the endpoint.

Tests updated to author class goals through the unified shape; the
legacy goals-PUT back-compat test is replaced by a whole-list-replace
test.
