+++ Sources/APIServer/Routes/Web/PublishedAssignmentRoutes+Achievements.swift
@@ -1,13 +1,14 @@
 // APIServer/Routes/Web/PublishedAssignmentRoutes+Achievements.swift
 //
-// Instructor authoring for class-wide goals (a kind of Achievement).
+// Instructor authoring for the unified Achievements editor table.
 //
-// GET returns the assignment's current class goals as editable rows; PUT
-// replaces the class-goal set in the manifest.  Class goals are
-// server-evaluated and display-only — they never change what the suite grades —
-// so, unlike suite edits, this path does NOT retest submissions, reschedule
-// validation, or close the assignment.  It just rewrites the manifest's
-// `achievements` array (preserving any non-class-goal achievements).
+// GET returns one editable row per achievement of any kind; PUT replaces the
+// manifest's `achievements` array wholesale.  `AchievementInput` is a flat,
+// editor-friendly projection of `Achievement`: the kind-specific parameters
+// appear as optional fields.  Achievements are server-evaluated and
+// display-only — they never change what the suite grades — so, unlike suite
+// edits, this path does NOT retest submissions, reschedule validation, or
+// close the assignment.
 
 import Core
 import Fluent
@@ -16,27 +17,38 @@ import Vapor
 
 extension PublishedAssignmentRoutes {
 
-    /// One editable class-goal row from the editor.
-    struct ClassGoalInput: Content {
+    /// One row in the unified Achievements table — any kind.
+    struct AchievementInput: Content {
         var id: String?
         var name: String
-        /// Per-student grade threshold, 0–100 (% of the assignment grade).
-        var thresholdPercent: Double
-        /// Share of the class that must reach the threshold, 0–100.
-        var classPercent: Double
-        /// Positive bonus points (extra credit, capped at 100% at grading time).
-        var bonusPoints: Int
+        /// Raw `AchievementKind` value (classGoal / thresholdBadge / testBadge /
+        /// firstTryPerfect / comeback / persistence / speedRun / classRecord).
+        var kind: String
+        var detail: String?
+        /// Grade cutoff %, 0–100 (classGoal + thresholdBadge target, and the
+        /// grade gate for firstTryPerfect / persistence / speedRun).
+        var thresholdPercent: Double?
+        /// classGoal: share of the class, 0–100.
+        var classPercent: Double?
+        /// classGoal: positive bonus points.
+        var points: Int?
+        /// testBadge: the test filename that must pass.
+        var testName: String?
+        /// classRecord: raw `RecordDimension`.
+        var recordDimension: String?
+        /// firstTryPerfect (max) / persistence (min) attempt cutoff.
+        var attemptThreshold: Int?
+        /// speedRun: total-execution cutoff in ms.
+        var timeThresholdMs: Int?
+        /// comeback: minimum grade jump in percentage points.
+        var jumpThresholdPercent: Int?
     }
 
     struct AchievementsBody: Content {
-        /// Legacy class-goals-card shape (replaces only the class-goal subset).
-        var goals: [ClassGoalInput]?
-        /// Unified shape (replaces the entire achievements list).  Wins when present.
-        var achievements: [AchievementInput]?
+        var achievements: [AchievementInput]
     }
 
     struct AchievementsResponse: Content {
-        var goals: [ClassGoalInput]
         var achievements: [AchievementInput]
     }
 
@@ -45,13 +57,7 @@ extension PublishedAssignmentRoutes {
     @Sendable
     func getAchievements(req: Request) async throws -> AchievementsResponse {
         let (_, setup) = try await loadAssignmentAndSetup(req)
-        return makeAchievementsResponse(fromManifest: setup.manifest)
-    }
-
-    private func makeAchievementsResponse(fromManifest manifest: String) -> AchievementsResponse {
-        AchievementsResponse(
-            goals: classGoalRows(fromManifest: manifest),
-            achievements: unifiedAchievementRows(fromManifest: manifest))
+        return AchievementsResponse(achievements: unifiedAchievementRows(fromManifest: setup.manifest))
     }
 
     // MARK: - PUT /instructor/:assignmentID/achievements
@@ -60,74 +66,146 @@ extension PublishedAssignmentRoutes {
     func putAchievements(req: Request) async throws -> AchievementsResponse {
         let (_, setup) = try await loadAssignmentAndSetup(req)
         let body = try req.content.decode(AchievementsBody.self)
-
-        let merged: [Achievement]
-        let isUnified = body.achievements != nil
-        if let unified = body.achievements {
-            // Unified shape: replace the entire achievements list.
-            merged = try unified.map { try achievement(fromUnified: $0) }
-        } else {
-            // Legacy class-goals shape: replace only the class-goal subset.
-            let newGoals = try (body.goals ?? []).map { try achievement(from: $0) }
-            let existing =
-                (try? JSONDecoder().decode(TestProperties.self, from: Data(setup.manifest.utf8)))?
-                .achievements ?? []
-            merged = existing.filter { $0.kind != .classGoal } + newGoals
-        }
+        let achievements = try body.achievements.map { try achievement(fromUnified: $0) }
 
         try await mutateManifest(setup: setup, on: req.db) { dict in
             let encoder = JSONEncoder()
             encoder.outputFormatting = [.sortedKeys]
-            dict["achievements"] = try JSONSerialization.jsonObject(with: encoder.encode(merged))
-            // A unified save means the instructor curated the full list — the
+            dict["achievements"] = try JSONSerialization.jsonObject(with: encoder.encode(achievements))
+            // Saving the table means the instructor curated the full list — the
             // manifest is now authoritative (built-in defaults no longer merge in).
-            if isUnified { dict["builtInAchievementsSeeded"] = true }
+            dict["builtInAchievementsSeeded"] = true
         }
-        return makeAchievementsResponse(fromManifest: setup.manifest)
+        return AchievementsResponse(achievements: unifiedAchievementRows(fromManifest: setup.manifest))
     }
 
-    // MARK: - Conversions
+    // MARK: - Row ↔ domain conversions
 
-    private func achievement(from input: ClassGoalInput) throws -> Achievement {
-        let trimmed = input.name.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty else {
-            throw WebAssignmentError.invalidParameter(name: "name", reason: "Goal name must not be empty.")
+    func achievement(fromUnified input: AchievementInput) throws -> Achievement {
+        let name = input.name.trimmingCharacters(in: .whitespacesAndNewlines)
+        guard !name.isEmpty else {
+            throw WebAssignmentError.invalidParameter(name: "name", reason: "Name must not be empty.")
         }
-        guard (0...100).contains(input.thresholdPercent),
-            (0...100).contains(input.classPercent),
-            input.bonusPoints >= 0
-        else {
+        guard let kind = AchievementKind(rawValue: input.kind) else {
             throw WebAssignmentError.invalidParameter(
-                name: "goal",
-                reason: "Threshold and class share must be 0–100; bonus points must be ≥ 0.")
+                name: "kind", reason: "Unknown achievement kind '\(input.kind)'.")
         }
-        let goalID: String
-        if let existing = input.id, !existing.isEmpty {
-            goalID = existing
-        } else {
-            goalID = "goal_\(UUID().uuidString.prefix(8))"
+        let id = unifiedAchievementID(input.id)
+        let detail = input.detail?.trimmingCharacters(in: .whitespacesAndNewlines)
+        let cleanDetail = (detail?.isEmpty == false) ? detail : nil
+
+        switch kind {
+        case .classGoal:
+            return try classGoalAchievement(input, id: id, name: name, detail: cleanDetail)
+        case .thresholdBadge:
+            return Achievement(
+                id: id, name: name, detail: cleanDetail, kind: .thresholdBadge, scope: .individual,
+                reward: AchievementReward(type: .badge, label: name),
+                target: AchievementTarget(kind: .assignmentGrade),
+                threshold: try gradeFraction(input.thresholdPercent, fallback: 0.9))
+        case .testBadge:
+            let test = (input.testName ?? "").trimmingCharacters(in: .whitespacesAndNewlines)
+            guard !test.isEmpty else {
+                throw WebAssignmentError.invalidParameter(
+                    name: "testName", reason: "A test badge needs the filename of the test that must pass.")
+            }
+            return Achievement(
+                id: id, name: name, detail: cleanDetail, kind: .testBadge, scope: .individual,
+                reward: AchievementReward(type: .badge, label: name),
+                target: AchievementTarget(kind: .testPass, ref: test))
+        case .firstTryPerfect, .persistence:
+            return Achievement(
+                id: id, name: name, detail: cleanDetail, kind: kind, scope: .individual,
+                reward: AchievementReward(type: .badge, label: name),
+                target: AchievementTarget(kind: .assignmentGrade),
+                threshold: try gradeFraction(input.thresholdPercent, fallback: 1.0),
+                attemptThreshold: input.attemptThreshold ?? (kind == .firstTryPerfect ? 1 : 5))
+        case .comeback:
+            return Achievement(
+                id: id, name: name, detail: cleanDetail, kind: .comeback, scope: .individual,
+                reward: AchievementReward(type: .badge, label: name),
+                target: AchievementTarget(kind: .assignmentGrade),
+                jumpThresholdPercent: input.jumpThresholdPercent ?? 50)
+        case .speedRun:
+            return Achievement(
+                id: id, name: name, detail: cleanDetail, kind: .speedRun, scope: .individual,
+                reward: AchievementReward(type: .badge, label: name),
+                target: AchievementTarget(kind: .assignmentGrade),
+                threshold: try gradeFraction(input.thresholdPercent, fallback: 1.0),
+                timeThresholdMs: input.timeThresholdMs ?? 2_000)
+        case .classRecord:
+            guard let dim = RecordDimension(rawValue: input.recordDimension ?? "firstToSolve") else {
+                throw WebAssignmentError.invalidParameter(
+                    name: "recordDimension", reason: "Unknown record dimension '\(input.recordDimension ?? "")'.")
+            }
+            return Achievement(
+                id: id, name: name, detail: cleanDetail, kind: .classRecord, scope: .classWide,
+                reward: AchievementReward(type: .title, label: name), recordDimension: dim)
+        }
+    }
+
+    private func classGoalAchievement(
+        _ input: AchievementInput, id: String, name: String, detail: String?
+    ) throws -> Achievement {
+        guard let classPercent = input.classPercent, (0...100).contains(classPercent) else {
+            throw WebAssignmentError.invalidParameter(
+                name: "classPercent", reason: "Class share must be 0–100.")
+        }
+        let points = input.points ?? 0
+        guard points >= 0 else {
+            throw WebAssignmentError.invalidParameter(name: "points", reason: "Bonus points must be ≥ 0.")
         }
         return Achievement(
-            id: goalID,
-            name: trimmed,
-            kind: .classGoal,
-            scope: .classWide,
-            reward: AchievementReward(type: .points, label: trimmed, points: input.bonusPoints),
+            id: id, name: name, detail: detail, kind: .classGoal, scope: .classWide,
+            reward: AchievementReward(type: .points, label: name, points: points),
             target: AchievementTarget(kind: .assignmentGrade),
-            threshold: input.thresholdPercent / 100,
-            classFraction: input.classPercent / 100)
+            threshold: try gradeFraction(input.thresholdPercent, fallback: 0),
+            classFraction: classPercent / 100)
     }
 
-    private func classGoalRows(fromManifest manifest: String) -> [ClassGoalInput] {
-        guard let props = try? JSONDecoder().decode(TestProperties.self, from: Data(manifest.utf8))
-        else { return [] }
-        return props.achievements.filter { $0.kind == .classGoal }.map { a in
-            ClassGoalInput(
-                id: a.id,
-                name: a.name,
-                thresholdPercent: (a.threshold ?? 0) * 100,
-                classPercent: (a.classFraction ?? 0) * 100,
-                bonusPoints: a.reward.points ?? 0)
+    /// Validates a 0–100 grade % and returns the 0…1 fraction; `fallback` (a
+    /// fraction) is used when the field is absent.
+    private func gradeFraction(_ percent: Double?, fallback: Double) throws -> Double {
+        let pct = percent ?? (fallback * 100)
+        guard (0...100).contains(pct) else {
+            throw WebAssignmentError.invalidParameter(
+                name: "thresholdPercent", reason: "Grade threshold must be 0–100.")
         }
+        return pct / 100
+    }
+
+    private func unifiedAchievementID(_ id: String?) -> String {
+        if let id, !id.isEmpty { return id }
+        return "ach_\(UUID().uuidString.prefix(8))"
+    }
+
+    func achievementInput(from a: Achievement) -> AchievementInput {
+        AchievementInput(
+            id: a.id, name: a.name, kind: a.kind.rawValue, detail: a.detail,
+            thresholdPercent: a.threshold.map { $0 * 100 },
+            classPercent: a.classFraction.map { $0 * 100 },
+            points: a.reward.points,
+            testName: a.target.kind == .testPass ? a.target.ref : nil,
+            recordDimension: a.recordDimension?.rawValue,
+            attemptThreshold: a.attemptThreshold,
+            timeThresholdMs: a.timeThresholdMs,
+            jumpThresholdPercent: a.jumpThresholdPercent)
+    }
+
+    /// The rows shown in the unified editor table.  Until the instructor first
+    /// saves the table (`builtInAchievementsSeeded`), the built-in defaults the
+    /// manifest hasn't authored yet are merged in so they appear as editable
+    /// rows; after seeding, the manifest is authoritative (a removed built-in
+    /// stays removed).
+    func unifiedAchievementRows(fromManifest manifest: String) -> [AchievementInput] {
+        guard let props = try? JSONDecoder().decode(TestProperties.self, from: Data(manifest.utf8))
+        else { return BuiltInAchievements.all.map(achievementInput(from:)) }
+        let authored = props.achievements
+        guard !props.builtInAchievementsSeeded else {
+            return authored.map(achievementInput(from:))
+        }
+        let authoredIDs = Set(authored.map(\.id))
+        let defaults = BuiltInAchievements.all.filter { !authoredIDs.contains($0.id) }
+        return (authored + defaults).map(achievementInput(from:))
     }
 }
